Added Save Functionality and MenuBar
Moved functional tiles that added tiles and folders to a menu item and removed them as tiles. Added save functionality, this is implimented through autosave. meaning there is no 'save' button but rather all changes are automatically written to the current profile. Note: there is a bug with this save/load, if a file with no children is saved then it will be loaded as a talker tile not a file, this can be fixed in a future patch.