Added ability to speak in the History Window
I edited the WindowHistory.xaml to have Buttons in the scroll frame. These Buttons all have a TextBlock as a child. When one of the phrases is clicked it is sent to an instance variable that is able to be accessed by the rest of the class. The currently selected button changes its color to reflect that it has been selected. If you click on the currently selected button again it will effectively clear the selected phrase.

I also migrated all of the TextBlock styling to a style tag at the top of the file.