Added some navigation functionality
All of the Windows have been converted to extend the class TalkerView. TalkerView extends the UserControl class which is a built in C# class. This allows us to have all of the functionality associated with UserControl and our own functionality that we define in TalkerView. Views were refactored to be part of their own folder.