Tokenize strings when spoken
TextUtility will now tokenize the spoken strings and store each word in a map as a key-value pair with the word being the key and the total number of times it was ever spoken as the value. Currently words that are the same but have different cases are treated the same i.e. "hello" and "HELLO" are considered to be equivalent.