Merge Current Utilities into Dev (#2)
* Added Utilities Folder

This will store all of the classes that will help the behind the scenes magic happen

* Added the TextUtility class

This is a singleton class that saves/loads the spoken phrases.

* Pushing to show to ppl, autoscan class
Not complete

* mainwindow startautoscan is now toggle autoscan

* autoscan starts on every page by default

* autoscan now works on keyboard page

* Tokenize strings when spoken

TextUtility will now tokenize the spoken strings and store each word in a map as a key-value pair with the word being the key and the total number of times it was ever spoken as the value. Currently words that are the same but have different cases are treated the same i.e. "hello" and "HELLO" are considered to be equivalent.