Code Cleanup of the Phrases Page
Added a new popup window called KeyboardPopup. This will allow us to make a keyboard on the fly. Whenever the user needs to type something just create a new KeyboardPopup object and call GetUserInput. This will return the typed string that the user types.

Also attempted to simplify the logic for the ImageGenerator class. This new logic is sitting in a class called ImageGeneratorTest.